ai_service/requirements.txt
# FastAPI AI Service
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
faiss-cpu==1.7.4
numpy==1.24.3
Pillow==10.1.0

# High-accuracy recognition (ArcFace / InsightFace)
insightface==0.7.3
onnxruntime==1.17.0
